+++ pcs/cli/resource_agent.py
@@ -3,16 +3,14 @@ from typing import List
 from pcs.cli.common.errors import CmdLineInputError
 from pcs.cli.common.tools import print_to_stderr
 from pcs.common import reports
-from pcs.common.resource_agent.dto import ResourceAgentNameDto
+from pcs.common.resource_agent.dto import (
+    ResourceAgentNameDto,
+    get_resource_agent_full_name,
+)
 
 
 def is_stonith(agent_name: ResourceAgentNameDto) -> bool:
     return agent_name.standard == "stonith"
-
-
-def get_resource_agent_full_name(agent_name: ResourceAgentNameDto) -> str:
-    provider = f":{agent_name.provider}" if agent_name.provider else ""
-    return f"{agent_name.standard}{provider}:{agent_name.type}"
 
 
 def find_single_agent(